Move punch::Player to a separate module
