+++ src/engine/punch/mod.rs
@@ -1,4 +1,6 @@
 pub use self::scene::Scene;
+pub use self::player::Player;
 
 pub mod scripting;
+mod player;
 mod scene;
